Move test_env() to test_unpackers_common.py
